+++ src/lib/CompendiumHelper.js
@@ -192,7 +192,8 @@ const CompendiumHelper = {
    * @param  {string} packageType - package type of compendium, defaults to world
    * @returns {object} - Object consisting of compendium and creation result
    */
-  createIfNotExists: async ({ label, type, id = undefined, packageType = "world", folderId = null } = {}) => {
+  // eslint-disable-next-line no-unused-vars
+  createIfNotExists: async ({ label, type, id = undefined, packageType = "world", folderId = null, dnd5eTypeTags = [] } = {}) => {
     if (id) logger.debug(`Checking if Compendium with id ${id} exists for ${SETTINGS.MODULE_ID}`);
     else if (label) logger.debug(`Checking if Compendium with label ${label} exists for ${SETTINGS.MODULE_ID}`);
     const compendium = await game.packs.get(id) ?? game.packs.find((p) => p.metadata.label === label);
@@ -219,6 +220,11 @@ const CompendiumHelper = {
           label,
           name,
           package: packageType,
+          // flags: {
+          //   dnd5e: {
+          //     types: dnd5eTypeTags,
+          //   }
+          // }
         });
         if (folderId) await newCompendium.setFolder(folderId);
         return {